do much

+++ src/Htl/SpendenportalBundle/Controller/ProjectController.php
@@ -634,12 +634,12 @@ class ProjectController extends Controller
                 $target_file = $target_dir . basename($_FILES["titlePictureUrl"]["name"]);
                 $uploadOk = 1;
 
-                //if(is_null($project->getTitlePictureUrl())){
+                if($project->getTitlePictureUrl() == ""){
                     if(unlink($target_dir.$project->getTitlePictureUrl())){
                         $uploadOk = 1;
                     }
-                //}
-                $project->setTitlePictureUrl($data["titlePictureUrl"]);
+                    $project->setTitlePictureUrl($data["titlePictureUrl"]);
+                }
 
                 $imageFileType = pathinfo($target_file, PATHINFO_EXTENSION);
                 // Check if image file is a actual image or fake image
@@ -710,16 +710,12 @@ class ProjectController extends Controller
                 );
             }
 
-            $projects = $this->getDoctrine()->getRepository('HtlSpendenportalBundle:User')->find(5)->getProjects();
+            $project->setDeleted(true);
 
-            //if($this->getUser()->getProjects()->find($projectId)){
-             if($projects->find($projectId)){
-                $project->setDeleted(true);
+            $em->flush();
 
-                $em->flush();
+            return new JsonResponse('Project has been deleted!');
 
-                return new JsonResponse('Project has been deleted!');                
-            }
             
             //$em->remove($project);
